requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-dotenv>=1.0.1
httpx>=0.27.0
requests>=2.32.3
pydantic>=2.7.4
pydantic-settings>=2.3.4
playwright>=1.44.0
pandas>=2.2.2
openpyxl>=3.1.2
PyPDF2>=3.0.1
pdfplumber>=0.11.0
pillow>=10.3.0
numpy>=1.26.4
matplotlib>=3.8.4
seaborn>=0.13.2
plotly>=5.22.0
kaleido>=0.2.1
openai>=1.30.1
tenacity>=8.3.0
pytest>=8.3.0
